% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_helpers.R
\name{create_config_table}
\alias{create_config_table}
\title{Title}
\usage{
create_config_table(
  ad_in,
  measures,
  diffs,
  dimreds,
  default_factors,
  db_prefix = "test1",
  db_dir = "data-raw"
)
}
\arguments{
\item{db_dir}{}
}
\value{

}
\description{
Title
}
\examples{
 TODO
}
