% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_additional_info.R
\name{display_document}
\alias{display_document}
\title{Display a document in Shiny.}
\usage{
display_document(file)
}
\arguments{
\item{file}{The file (path) containing a document to display.}
}
\value{
html tags of the rendered doc
}
\description{
This function will take a Markdown document (.md) or Word document (.docx) and
return the UI for a Shiny app to display it. This should be embedded within a
UI code (see examples for details).
}
\examples{

\dontrun{
ui <- shiny::fluidPage(
shiny::mainPanel(
display_document('path-to-your-file')
)
)
}
}
