% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pg_vis_comp.R
\name{do_stat_test}
\alias{do_stat_test}
\title{Do a statitical test}
\usage{
do_stat_test(
  in_meta,
  in_data,
  group,
  group1_name,
  group2_name,
  colorby_group,
  normalization = c("raw", "tot_area"),
  transformation = c("none", "log10"),
  test
)
}
\arguments{
\item{in_data}{tibble in tidy format}

\item{group}{what column contains the group info}

\item{group1_name}{is the name of group 1}

\item{group2_name}{is the name of group 2}

\item{normalization}{what normalization to use, none (raw data) or total area normalization}

\item{transformation}{what transformation to use}

\item{test}{do a t-test or Mann-Whitney U test}
}
\value{
a tibble ready for statistical testing
}
\description{
Do a statitical test, t-test or Mann-Whitney U test
}
\author{
Rico Derks
}
