% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{get_config}
\alias{get_config}
\title{Helper for reading the Config file containing information about the databases}
\usage{
get_config(in_path = NULL)
}
\arguments{
\item{in_path}{where the config file lives. default getwd() test}
}
\value{
the list of options contained in `app_config.yml`
}
\description{
Helper for reading the Config file containing information about the databases
}
\examples{
# TODO

}
