% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{get_db_conf}
\alias{get_db_conf}
\title{get_db_conf: helper for reading the database config}
\usage{
get_db_conf(db_name, db_root = NULL)
}
\arguments{
\item{db_name}{what is out db called}

\item{db_root}{where the folder lives. default cwd}
}
\value{
config_list: list of UI configurations for accessing data
}
\description{
get_db_conf: helper for reading the database config
}
\examples{
TODO
}
