% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pg_expression.R
\name{get_my_cols2}
\alias{get_my_cols2}
\title{Get colors for the top annotations}
\usage{
get_my_cols2(top_annotations)
}
\arguments{
\item{top_annotations}{top annotations}
}
\value{
colors
}
\description{
Get colors for the top annotations
}
