% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pg_diff_expr.R
\name{hline}
\alias{hline}
\title{Create horizontal line in plotly plot}
\usage{
hline(y = 0, color = "blue")
}
\arguments{
\item{y}{at what position on the y-axis}

\item{color}{color of the line}
}
\value{
Returns a list with all the settigns needed
}
\description{
Create horizontal line in plotly plot
}
