% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pg_expression.R
\name{make_cx_heatmap}
\alias{make_cx_heatmap}
\title{Make complex heatmap}
\usage{
make_cx_heatmap(
  in_mat,
  cluster_samps,
  samp_grp,
  samp_grp_nm,
  samp_title,
  samp_aggregated,
  samp_split,
  cluster_feats,
  feat_grp,
  feat_grp_nm,
  feats_title,
  units_label,
  omics,
  omics_at,
  top_annotations,
  right_annotations
)
}
\arguments{
\item{in_mat}{input matrix}

\item{cluster_samps}{cluster the samples}

\item{samp_grp}{group the samples}

\item{samp_grp_nm}{group names of the samples}

\item{samp_title}{title of the samples}

\item{samp_aggregated}{aggregate the samples}

\item{samp_split}{split the samples}

\item{cluster_feats}{cluster features}

\item{feat_grp}{group the features}

\item{feat_grp_nm}{group names of the features}

\item{feats_title}{title of the features}

\item{units_label}{label of the units}

\item{omics}{omics}

\item{omics_at}{omics at}

\item{top_annotations}{show top annotations}

\item{right_annotations}{right annotations}
}
\value{
Complex heatmap object
}
\description{
Make complex heatmap
}
