% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_helpers.R
\name{pack_anndata_from_csv}
\alias{pack_anndata_from_csv}
\title{pack_anndata}
\usage{
pack_anndata_from_csv(data_in)
}
\arguments{
\item{data_in}{- list of matrices, tables, and lists containing the data to pack into the db}
}
\value{
adata the anndata object we are browsing
}
\description{
pack_anndata
}
\examples{
# TODO
}
