% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_helpers.R
\name{pack_anndata_from_seurat}
\alias{pack_anndata_from_seurat}
\title{pack_anndata_from_seurat}
\usage{
pack_anndata_from_seurat(seurat_obj_name)
}
\arguments{
\item{seurat_obj_name}{name of seurat object}
}
\value{
adata the anndata object we are browsing
}
\description{
pack_anndata_from_seurat
}
