% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pg_vis_raw.R
\name{pg_violin_box}
\alias{pg_violin_box}
\title{pg_violin_box
pg_violin_box
#' pg_violin_box}
\usage{
pg_violin_box(
  in_conf,
  in_meta,
  in_fact,
  in_quant,
  in_grp,
  in_subset,
  in_subsel,
  in_data,
  in_omic,
  plot_type,
  show_data_points,
  data_point_sz,
  font_size
)
}
\arguments{
\item{in_conf}{config}

\item{in_meta}{meta}

\item{in_fact}{x-axis grouping}

\item{in_quant}{y-axis quantity to visualize}

\item{in_grp}{grouping variable}

\item{in_subsel}{subset of group}

\item{in_data}{raw_data (in_quant values)}

\item{in_omic}{omic for subset (Not implimented)}

\item{plot_type}{violin or box}

\item{show_data_points}{boolean}

\item{data_point_sz}{size - smaller if too many data points is better}

\item{font_size}{font size}
}
\value{
rendered ggplot
}
\description{
playground bubbleplot / heatmap function on a ggplot backbone
}
\examples{
TBD
}
