% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pg_vis_comp.R
\name{pg_volcano_ly}
\alias{pg_volcano_ly}
\title{Create volcano plot}
\usage{
pg_volcano_ly(in_data, pvalue_adjust = FALSE, title = "")
}
\arguments{
\item{in_data}{tibble with all the lipid data and test data}

\item{pvalue_adjust}{show the corrected p value, default is FALSE}

\item{title}{title of the plot}
}
\value{
plotly object
}
\description{
Create volcano plot.
}
\author{
Andy Henrie modified from Rico Derks
}
