% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_defaults}
\alias{run_defaults}
\title{run_defaults - shortcut to launch with default settings.}
\usage{
run_defaults()
}
\description{
run_defaults - shortcut to launch with default settings.
}
\examples{
# TODO
}
