% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_docker}
\alias{run_docker}
\title{Shortcut to launch with default settings for the docker container}
\usage{
run_docker()
}
\description{
Shortcut to launch with default settings for the docker container
}
