% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_in_browser}
\alias{run_in_browser}
\title{run_in_browser - shortcut to launch in_browser}
\usage{
run_in_browser(
  db_root = NULL,
  database_names = NULL,
  install_type = "empty",
  ...
)
}
\arguments{
\item{db_root}{path of the database root}

\item{database_names}{database names}

\item{install_type}{installation type}

\item{...}{parameters passes on to run_app()}
}
\description{
run_in_browser - shortcut to launch in_browser
}
\examples{
# TODO
}
