% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_helpers.R
\name{setup_database}
\alias{setup_database}
\title{setup_database}
\usage{
setup_database(database_name, db_path, data_in, db_meta, re_pack = TRUE)
}
\arguments{
\item{re_pack}{}
}
\value{
The return value, if any, from executing the function.
}
\description{
A a function to create the anndata database
}
\examples{
 TODO
}
