% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pg_diff_expr.R
\name{vline}
\alias{vline}
\title{Create vertical line in plotly plot}
\usage{
vline(x = 0, color = "blue")
}
\arguments{
\item{x}{at what position on the x-axis}

\item{color}{color of the line}
}
\value{
Returns a list with all the settigns needed
}
\description{
Create vertical line in plotly plot
}
