% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_additional_info.R
\name{word_to_markdown}
\alias{word_to_markdown}
\title{Convert a Word (docx) to a Markdown (md) file}
\usage{
word_to_markdown(file, output = NULL)
}
\arguments{
\item{file}{The path to your .docx file.}

\item{output}{The path to save your new markdown file}
}
\value{
The same file as a Markdown document.
}
\description{
This function converts a word document to a markdown file.
If output path is NULL (as default), the function will save this to a temporary path.
}
\examples{

#ADD LATER

}
