% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{write_db_conf}
\alias{write_db_conf}
\title{write_db_conf: helper for writing the ui config options for the database}
\usage{
write_db_conf(config_list, db_name, db_root = NULL)
}
\arguments{
\item{config_list}{the database configuration list}

\item{db_name}{what is out database is called}

\item{db_root}{where the folder lives. default cwd}
}
\value{
Nothing, a file is written.
}
\description{
write_db_conf: helper for writing the ui config options for the database
}
\examples{
# TODO
}
