% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{write_db_meta}
\alias{write_db_meta}
\title{write_db_meta: helper for writing the database meta info}
\usage{
write_db_meta(db_meta, db_name, db_root = NULL)
}
\arguments{
\item{db_meta}{data to write}

\item{db_name}{what is out database is called}

\item{db_root}{where the folder lives. default cwd}
}
\value{
Nothing, a file is written.
}
\description{
write_db_meta: helper for writing the database meta info
}
\examples{
# TODO
}
