% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_helpers.R
\name{compute_de_table}
\alias{compute_de_table}
\title{Title}
\usage{
compute_de_table(adata, comp_types, test_types, obs_names, sc)
}
\arguments{
\item{adata}{the anndata object}

\item{comp_types}{what kind of comparisons?  "allVrest" or "{a}V{b}"}

\item{test_types}{statistical tests - t-test, etc}

\item{obs_names}{name of the adata$obs column defining the comparision groups}

\item{sc}{scanpy}
}
\value{

}
\description{
Title
}
\examples{
TODO
}
