% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{get_config}
\alias{get_config}
\title{get_config: helper for reading the Config file containing information about the databases}
\usage{
get_config(in_path = NULL, is_running = FALSE)
}
\arguments{
\item{in_path}{where the config file lives. default getwd()}
}
\value{

}
\description{
get_config: helper for reading the Config file containing information about the databases
}
\examples{
TODO
}
