% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{get_db_meta}
\alias{get_db_meta}
\title{Helper for reading the database meta data}
\usage{
get_db_meta(db_name, db_root = NULL)
}
\arguments{
\item{db_name}{what is out db called}

\item{db_root}{where the folder lives. default cwd}
}
\value{
meta_info: database meta deta
}
\description{
Helper for reading the database meta data
}
\examples{
# TODO

}
