% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_in_browser}
\alias{run_in_browser}
\title{Shortcut to launch in_browser}
\usage{
run_in_browser(
  db_root = NULL,
  database_names = NULL,
  install_type = "empty",
  ...
)
}
\arguments{
\item{db_root}{root path of the database}

\item{database_names}{names of all databases}

\item{install_type}{install type}

\item{...}{parameters to pass to `run_app`}
}
\description{
Run the app in the browser instead of Rstudio when started
    from Rstudio.
}
