% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{write_config}
\alias{write_config}
\title{write_config: helper for Writing the Config file for setting up the omicser databases}
\usage{
write_config(in_options, in_path = NULL)
}
\arguments{
\item{in_path}{where the config file lives. default cwd}
}
\value{

}
\description{
write_config: helper for Writing the Config file for setting up the omicser databases
}
\examples{
TODO
}
