% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{write_config}
\alias{write_config}
\title{Helper for Writing the Config file for setting up the omicser databases}
\usage{
write_config(in_config, in_path = NULL, set_default = FALSE)
}
\arguments{
\item{in_config}{list of database names, conda environment, and database path}

\item{in_path}{where the config file lives. default cwd}

\item{set_default}{set default}
}
\value{
Nothing returned, writes to file.
}
\description{
Helper for Writing the Config file for setting up the omicser databases
}
\examples{
# TODO

}
